fix: tcpassembly does not use correct timestamps
tcpassembly only read the bytes of the reassembled stream and ignored
other reassembly information, like seen timestamp. Therefore When
assembling a HEP packet it had to use the current time. Because of
parallel handling of packet capture, reassembly, and extraction this
timestamp could be totally wrong.

The changed code processes the reassembly information directly and
uses the timestamp of the part with the begin of the SIP message.

refs: https://github.com/sipcapture/heplify/issues/223